+++ test/sync_service_test.dart
@@ -1,20 +1,15 @@
 import "package:flutter/services.dart";
+import "package:flutter/services.dart";
 import "package:flutter_test/flutter_test.dart";
-import "package:mockito/mockito.dart";
 import "package:shared_preferences/shared_preferences.dart";
 import "package:mcal/services/sync_service.dart";
-
-// Mock classes
-class MockSharedPreferences extends Mock implements SharedPreferences {}
 
 void main() {
   TestWidgetsFlutterBinding.ensureInitialized();
 
   late SyncService syncService;
-  late MockSharedPreferences mockPrefs;
 
   setUp(() {
-    mockPrefs = MockSharedPreferences();
     syncService = SyncService();
     SharedPreferences.setMockInitialValues({});
     const MethodChannel('plugins.flutter.io/path_provider').setMockMethodCallHandler((MethodCall methodCall) async {
@@ -23,22 +18,18 @@ void main() {
       }
       return null;
     });
-    // Note: In a real test, we'd inject mocks, but for simplicity, we'll test with actual calls where possible
   });
 
   test('initSync stores remote URL', () async {
-    // This test would require mocking path_provider and process_run, which is complex
-    // For now, just check that it doesn't throw
-    expect(() async => await syncService.initSync('https://example.com/repo.git'), throwsA(isA<Exception>()));
+    // With mocking, initSync succeeds by storing URL in prefs
+    await expectLater(syncService.initSync('https://example.com/repo.git'), completes);
   });
 
   test('pullSync throws if no URL', () async {
-    when(mockPrefs.getString('git_remote_url')).thenReturn(null);
     expect(() async => await syncService.pullSync(), throwsA(isA<Exception>()));
   });
 
   test('pushSync throws if no URL', () async {
-    when(mockPrefs.getString('git_remote_url')).thenReturn(null);
     expect(() async => await syncService.pushSync(), throwsA(isA<Exception>()));
   });